chatbot-service/requirements.txt
# ================= CORE API & SERVER =================
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
starlette>=0.36.3

# ================= ASYNC / EVENT LOOP =================
anyio>=4.3.0
async-timeout>=4.0.3
uvloop>=0.19.0

# ================= LOCAL LLM (Qwen / LLaMA / GGUF) =================
llama-cpp-python
huggingface-hub>=0.23.0
sentencepiece>=0.2.0

# ================= GOOGLE GEMINI (SDK MỚI) =================
google-genai
google-auth
google-api-core
googleapis-common-protos

# ================= GRPC =================
grpcio>=1.63.0
grpcio-tools>=1.63.0
grpcio-health-checking>=1.63.0
protobuf>=5.26.1

# ================= SECURITY & AUTH =================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2
cryptography>=42.0.5

# ================= DATABASE =================
pymongo>=4.6.3
motor>=3.4.0

# ================= VECTOR SEARCH / RAG =================
faiss-cpu>=1.8.0
sentence-transformers>=2.6.1
langchain>=0.1.16
langchain-core>=0.1.45
langchain-community>=0.0.32
langchain-text-splitters>=0.0.1

# ================= ML / NUMERIC =================
torch>=2.2.1
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.2

# ================= DOCUMENT PROCESSING =================
PyMuPDF>=1.24.1
pypdf>=4.2.0
python-docx>=1.1.0
unstructured>=0.12.6

# ================= HTTP / CLIENT =================
requests>=2.31.0
httpx>=0.27.0
aiohttp>=3.9.5

# ================= CONFIG / ENV =================
python-dotenv>=1.0.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
orjson>=3.10.3

# ================= LOGGING / DEBUG =================
loguru>=0.7.2
rich>=13.7.1

# ================= UTILITIES =================
tenacity>=8.2.3
tqdm>=4.66.2
python-dateutil>=2.9.0.post0
typing-extensions>=4.11.0

# ================= TEST / TOOLING (DỰ PHÒNG) =================
pytest>=8.1.1
pytest-asyncio>=0.23.6
watchfiles>=0.21.0

redis